Update quickstart.rst
This is a proposed example for the proj_create() function. This function has its own secularities, which can't be inferred from the example illustrating proj_create_crs_to_crs(). It is shown how to populate the longitude and latitude, how to convert to radians, transform to projected units, then transform back, and how to convert back to degrees. 

The code was compiled and tested. It is suggested to verify it compiles and prints what is expected.